fix(o11y): split storage gRPC tracing tests for process isolation
